requirements.txt
requests==2.25.1
django==3.2.5
psycopg2-binary==2.8.6
redis==3.5.3
celery==5.1.2
django-ckeditor==6.1.0
django-currentuser==0.5.3